Update meeting-search.php
This allows the pronunciation of the city to be changed. 
Note: unfortunately if you are using the sms function it alters the spelling of the city also.  Anyone with any ideas?
If the variables are not set in config.php or are hashed out then it doesn't affect anything.
added line:           
$results[2] = preg_replace($city_pron_fix_a, $city_pron_fix_b, $results[2]);
Explanation:
String variables $city_pron_fix_a and $city_pron_fix_b are set in config.php in the form of:
static $city_pron_fix_a = "/actual spelling/";
static $city_pron_fix_b = "spelling for best pronunciation";
Take for example Bossier:
static $city_pron_fix_a = "/Bossier/";
static $city_pron_fix_b = "Boes-yur";

Changes to allow for more information to be given before saying goodbye. .
#One line was added and one new variable was added, meeting _search_more_info .

       <?php echo word('meeting_search_more_info')?> #added this line
meeting_search_more_info is set in config.php in the form of:
static $meeting_search_more_info = "whatever you want to add";
If the variables are not set in config.php or are hashed out then it doesn't affect anything.