Separated out the comparison between local config and tomato into its own function
Moved the rest of the functionality to getHelplineBMLTRootServer

Added lat/long/search type arguments into getServiceBodies so that it can be passed along to getHelplineBMLTRootServer in the case of coming from a volunteer call.

In all cases where no arguments are passed to getHelplineBMLTRootServer, it behaves identical to original function in the master file.

No global settings edited either in session or otherwise

Tested with each combination of settings - works.